Extract Shopify route types
Move shared route types into a cycle-free module so handlers and interceptors can depend on the same contracts while preserving existing public re-exports.
