+++ packages/hydrogen/src/core/interceptors/graphiql.ts
@@ -1,5 +1,5 @@
 import type { StorefrontClient } from "../../client";
-import type { HydrogenRouteInterceptor } from "../handle-shopify-routes";
+import type { HydrogenRouteInterceptor } from "../route-types";
 import type { GraphiQLOptions } from "../types";
 
 const DEV_SCHEMA_FETCH_TIMEOUT_MS = 10_000;
