+++ packages/hydrogen/src/core/interceptors/proxy.ts
@@ -1,6 +1,6 @@
-import type { HydrogenRouteInterceptor, HydrogenRoutesOptions } from "../handle-shopify-routes";
 import { extractHeaders, REQUEST_GROUP_ID_HEADER } from "../headers";
 import { getLogger } from "../logging";
+import type { HydrogenRouteInterceptor, HydrogenRoutesOptions } from "../route-types";
 
 const PROXY_TIMEOUT_MS = 30_000;
