+++ packages/hydrogen/src/core/route-types.ts
@@ -1,0 +1,78 @@
+import type { StorefrontClient } from "../client";
+import type { ShopifyRequestContext } from "./headers";
+
+type Awaitable<T> = T | Promise<T>;
+
+export type ShopifyRouteSessionManager = {
+  getSessionOrigin(): Awaitable<string>;
+  getSessionItem(key: string): Awaitable<unknown>;
+  setSessionItem(key: string, value: unknown): Awaitable<void>;
+  removeSessionItem(key: string): Awaitable<void>;
+  commit?(): Awaitable<HeadersInit | void>;
+};
+
+export type ShopifyRouteHandlerContext = {
+  request: Request;
+  sessionManager: ShopifyRouteSessionManager;
+  storefrontClient: StorefrontClient;
+  requestContext: ShopifyRequestContext;
+};
+
+export type ShopifyRouteJsonResult<TData = unknown> = {
+  type: "json";
+  data: TData;
+  headers?: HeadersInit;
+};
+
+export type ShopifyRouteRedirectResult = {
+  type: "redirect";
+  location: string;
+  headers?: HeadersInit;
+};
+
+export type ShopifyRouteError = {
+  code: string;
+  message: string;
+};
+
+export type ShopifyRouteErrorResult<TError extends ShopifyRouteError = ShopifyRouteError> = {
+  type: "error";
+  error: TError;
+  status?: number;
+  headers?: HeadersInit;
+};
+
+export type ShopifyRouteHandlerResult<
+  TData = unknown,
+  TError extends ShopifyRouteError = ShopifyRouteError,
+> = ShopifyRouteJsonResult<TData> | ShopifyRouteRedirectResult | ShopifyRouteErrorResult<TError>;
+
+export type CallableRouteHandler<
+  TContext,
+  TResult,
+  TPathname extends string = string,
+  TMethod extends string = string,
+> = ((context: TContext) => Promise<TResult>) & {
+  readonly pathname: TPathname;
+  readonly method: TMethod;
+};
+
+export type ShopifyRouteHandler<
+  TPathname extends string = string,
+  TMethod extends string = string,
+> = CallableRouteHandler<ShopifyRouteHandlerContext, ShopifyRouteHandlerResult, TPathname, TMethod>;
+
+export type ShopifyRouteHandlerGroup = Record<string, ShopifyRouteHandler>;
+
+export type HydrogenRoutesOptions = ShopifyRouteHandlerContext & {
+  handlers?: readonly ShopifyRouteHandlerGroup[];
+};
+
+export type HydrogenRouteHandler<TExtraOptions extends object = object> = (
+  options: HydrogenRoutesOptions & TExtraOptions,
+) => null | Promise<Response>;
+
+export type HydrogenRouteInterceptor<TExtraOptions extends object = object> = (
+  url: URL,
+  ...args: Parameters<HydrogenRouteHandler<TExtraOptions>>
+) => ReturnType<HydrogenRouteHandler<TExtraOptions>>;
